+++ projectroles/templatetags/projectroles_tags.py
@@ -1,12 +1,60 @@
 from django import template
-from django.conf import settings
 import mistune
 
-from ..models import RoleAssignment, OMICS_CONSTANTS
-from ..plugins import get_backend_api
+from ..models import Project, RoleAssignment, OMICS_CONSTANTS
+
+
+# Local constants
+INDENT_PX = 25
 
 
 register = template.Library()
+
+
+@register.simple_tag
+def get_project_list(user, parent=None):
+    """Return flat project list for displaying in templates"""
+    project_list = []
+
+    if user.is_superuser:
+        project_list = Project.objects.filter(
+            parent=parent,
+            submit_status='OK').order_by('title')
+
+    elif not user.is_anonymous():
+        project_list = [
+            p for p in Project.objects.filter(
+                parent=parent,
+                submit_status='OK').order_by('title')
+            if p.has_role(user, include_children=True)]
+
+    def append_projects(project):
+        lst = [project]
+
+        for c in project.get_children():
+            if (user.is_superuser or
+                    c.has_role(user, include_children=True)):
+                lst += append_projects(c)
+
+        return lst
+
+    flat_list = []
+
+    for p in project_list:
+        flat_list += append_projects(p)
+
+    return flat_list
+
+
+@register.simple_tag
+def get_project_list_indent(project, list_parent):
+    """Return indent in pixels for project list"""
+    project_depth = project.get_depth()
+
+    if list_parent:
+        project_depth -= (list_parent.get_depth() + 1)
+
+    return project_depth * INDENT_PX
 
 
 @register.simple_tag
@@ -23,34 +71,8 @@ def get_plugin_info(plugin, pk):
 
 
 @register.simple_tag
-def get_latest_activity(project, timeline):
-    """Return datetime for latest project activity from Timeline"""
-    # TODO: Get the latest activity which 1) is OK and 2) has changed data
-    # TODO: (need to make some modifications to timeline for this)
-
-    timeline = get_backend_api('timeline_backend')
-
-    if not timeline:
-        return '<td colspan="4"></td>\n'
-
-    events = timeline.get_project_events(project, classified=False)
-
-    if events.count() > 0:
-        latest_event = events.order_by('-pk')[0]
-
-        ret = '<td style="white-space: nowrap;">{}</td>\n<td>' \
-            '<a href="mailto:{}">{}</a></td>\n<td>{}</td>\n'.format(
-                latest_event.get_timestamp().strftime('%Y-%m-%d %H:%M'),
-                latest_event.user.email,
-                latest_event.user.username,
-                timeline.get_event_description(latest_event))
-        return ret
-
-
-@register.simple_tag
 def get_description(project):
     """Return description, truncate if needed"""
-    # TODO: Can be removed once we add a smart overflow popup in CSS
     max_len = 128
     ret = project.description[:max_len]
 
@@ -70,7 +92,6 @@ def get_user_role_str(project, user):
         return ''
 
 
-# TODO: Could make this a filter too I guess
 @register.simple_tag
 def render_markdown(raw_markdown):
     return mistune.markdown(raw_markdown)